move into new file, and support 5DBase
